fix: use AWS_DEFAULT_REGION for slack notification.
- separating slack notification code from the processing code.
